+++ client.js
@@ -690,7 +690,7 @@ async function instanceManagement(slaveConfig, instance, instanceconfig, server,
 		if(!global.subscribedFiles) {
 			global.subscribedFiles = {};
 		}
-		if(pluginConfig.binary == "nodePackage" && pluginConfig.enabled){
+		if (pluginConfig.enabled) {
 			// require plugin class and execute it
 			let pluginClass = require(path.resolve(pluginConfig.pluginPath, "index"));
 			plugins[i] = new pluginClass(combinedConfig, async function(data, callback){