+++ sharedPlugins/globalChat/config.js
@@ -5,6 +5,5 @@ module.exports = {
 	// Name of package. For display somewhere I guess.
 	name: "globalChat",
 	version: "1.0.0",
-	binary: "nodePackage",
 	description: "Clusterio plugin to allow for chat between instances.",
 }